AI_Shopping_Helper/requirements.txt
# AI Shopping Helper - Requirements

# Core Framework (updated for Python 3.13 compatibility)
fastapi>=0.110.0,<1.0.0
# Avoid heavy/OS-specific extras to reduce wheel issues on Windows
uvicorn>=0.29.0,<1.0.0

# Database
# SQLite works out of the box; add SQLAlchemy and a Postgres driver for PostgreSQL support
SQLAlchemy>=2.0.30,<3
psycopg[binary]>=3.1,<4

# Web Scraping
requests>=2.32.0,<3
beautifulsoup4>=4.12.3,<5
selenium>=4.20.0,<5
scrapy>=2.12.0,<3

# Data Processing (wheels available for Python 3.13)
pandas>=2.2.3,<3
numpy>=2.1.3,<3

# Image Processing
opencv-python>=4.10.0,<5
Pillow>=10.3.0,<11
pytesseract>=0.3.11,<0.4

# AI/ML Libraries
scikit-learn>=1.6.0,<2
sentence-transformers>=2.7.0,<3
# ChromaDB is optional; enable by installing `chromadb` and setting USE_CHROMA=true

# Image→Product (CLIP) pipeline
# CPU-optimized: open_clip_torch with ViT-B/32 for i5 13th gen + 16GB RAM
torch>=2.2
torchvision>=0.17
torchaudio>=2.2
open-clip-torch>=2.20.0
faiss-cpu>=1.7
tqdm>=4.66

# HTTP Client
aiohttp>=3.10.5,<4

# Configuration
python-dotenv>=1.0.1,<2

# Development Tools
pytest>=8.2.0,<9
pytest-asyncio>=0.23.7,<0.24

# Logging
structlog>=24.1.0,<25

# Utilities
python-dateutil>=2.9.0.post0,<3

# Production Server
gunicorn>=21.2.0,<22

# Caching
redis>=5.0.4,<6

# Environment Variables
python-decouple>=3.8,<4
